pizza_agent_backend/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-cloud-bigquery==3.11.4 